Fixed button swipes
Had to correct the indexes and lengths to properly update them.
onSwipe was missing the index update, as if it was done by hand, hence it was breaking the system.
Had to fix the indexing for each card so both way can work together, as the indexes were passed as names and not numbers